[#24] refactor: 파일 업로드 관련 메소드를 재사용할 수 있도록 수정
